# perf(build): switch build optimization from `ReleaseSafe` to `ReleaseFast`